requirements.txt
streamlit==1.30.0
ultralytics==8.2.20
torch>=2.1.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
